refactor: 폴더구조 및 이름 변경, 링크추가
폴더구조 및 이름 변경 작업, 헤더에 검색페이지 링크 추가
