+++ src/components/Profile/Follow/UserListItem/ProfileUserListItem.module.scss
@@ -3,7 +3,7 @@
 .item {
   display: flex;
   align-items: center;
-  padding: 1rem;
+  padding: 1rem 0;
   .image {
     width: 5rem;
     height: 5rem;
