+++ src/styles/mixin.scss
@@ -49,7 +49,7 @@
 }
 
 @mixin smallButtonCS {
-  width: 8rem;
+  width: 9rem;
   height: 4rem;
   font-size: 1.6rem;
   align-items: center;
